+++ docs/jest-example/src/index.test.ts
@@ -81,5 +81,3 @@ run(
   },
   {api: api.jest}
 );
-
-