+++ src/index.ts
@@ -1,6 +1,9 @@
 import {IApiMapper, ITestConfig} from './models';
 import {TestRunner, TestSuiteDefinition} from './test-runner';
 
+export {ITestConfig, IApiMapper} from './models';
+export {TestSuiteDefinition} from './test-runner';
+
 export function run(testSuiteDefinition: TestSuiteDefinition, config: ITestConfig) {
   new TestRunner(testSuiteDefinition, config).run();
 }